refactor: 기존 네트워크 모듈 객체를 di 패키지로 옮겼습니다.
